--- crates/asm/subprotocols/upgrade/src/multisig/mod.rs
@@ -1,1 +1,0 @@
-pub mod authority;
